### 2022-01-14 #4067 @whedon
[PRE REVIEW]: pymdp: A Python library for active inference in discrete state spaces
**Submitting author:** @conorheins (<a href="http://orcid.org/0000-0002-5884-7728">Conor Heins</a>)
**Repository:** <a href="https://github.com/infer-actively/pymdp" target ="_blank">https://github.com/infer-actively/pymdp</a>
**Version:** v0.0.4
