### 2022-01-16 #4067 @emdupre
:wave: Hi @conorheins, and thank you for your submission to JOSS !

If you have any suggestions for potential reviewers, please let me know by listing their names or GitHub handles (without the @, so they don't receive a notification) here.
