### 2022-01-17 efa4faf
numpy subclass proposal
