+++ pymdp/types/distribution.py
@@ -2,30 +2,29 @@ import numpy as np
 
 
 class Distribution(np.ndarray):
-    def __new__(subtype, shape, dtype=float, buffer=None, offset=0, strides=None, order=None):
+    def __new__(subtype, shapes, dtype=float, buffer=None, offset=0, strides=None, order=None):
         # Check if we are dealing and a multi-factor distribution
-        dims = len(shape)
-        factors = 1 if dims < 2 else shape[0]
 
-        if factors > 2:
-            obj = super().__new__(subtype, shape[0], object, buffer, offset, strides, order)
+        n_factors = len(shapes) if isinstance(shapes[0], tuple) else 1
+
+        if n_factors > 1:
+            obj = super().__new__(subtype, n_factors, object, buffer, offset, strides, order)
             # Ceate an ndarray for each for factor
-            for factor in range(factors):
-                dist_shape = (shape[1], shape[2])
+            for factor in range(n_factors):
                 obj[factor] = super().__new__(
-                    subtype, dist_shape, dtype, buffer, offset, strides, order
+                    subtype, shapes[factor], dtype, buffer, offset, strides, order
                 )
 
         else:
-            obj = super().__new__(subtype, shape, dtype, buffer, offset, strides, order)
+            obj = super().__new__(subtype, shapes, dtype, buffer, offset, strides, order)
 
-        obj.factors = factors
+        obj.n_factors = n_factors
         return obj
 
     def __array_finalize__(self, obj):
         if obj is None:
             return
-        self.factors = getattr(obj, "factors", 1)
+        self.n_factors = getattr(obj, "n_factors", 1)
 
     def dot(self):
         # Can override behavior based on ufuncs (true for all maths functions)
@@ -33,9 +32,23 @@ class Distribution(np.ndarray):
 
 
 if __name__ == "__main__":
+    
+    # single factor example
+    dist_b = Distribution(shapes=(6, 2, 1))
+    print(f'Number of factors: {dist_b.n_factors}')
+    print(dist_b * 0)
+    print(dist_b + 10)
+    # print(dist_b[0] * 0)
+    # print(dist_b[0] + 10)
+    # print(np.dot(dist_b, dist_b))
 
-    dist_b = Distribution(shape=(6, 2, 1))
-    print(dist_b.factors)
-    print(dist_b[0] * 0)
-    print(dist_b[0] + 10)
-    print(np.dot(dist_b, dist_b))
+    # multi factor example
+    dist_c = Distribution(shapes = ((6, 2, 1), (5, 3, 3)))
+    print(f'Number of sub-arrays: {dist_c.n_factors}')
+    for factor in range(dist_c.n_factors):
+        print(f'Shape of sub-array {factor}: {dist_c[factor].shape}')
+
+    print('I don''t understand why the values of the arrays, are what they are though...?\n')
+    added_dist_c = dist_c + 10
+    print(added_dist_c[0])
+    print(added_dist_c[1])
