### 2022-01-19 1623215
- Added a "total_time" parameter for fixed horizon simulations - Allowed option to return more info from inferring policies - Added option for alpha parameter - Added expected utility & expected information trackers in control.py - Removed squeeze from probabilities.squeeze() in utils.py
