+++ inferactively_pymdp.egg-info/SOURCES.txt
@@ -1,0 +1,34 @@
+LICENSE
+README.md
+setup.py
+inferactively_pymdp.egg-info/PKG-INFO
+inferactively_pymdp.egg-info/SOURCES.txt
+inferactively_pymdp.egg-info/dependency_links.txt
+inferactively_pymdp.egg-info/requires.txt
+inferactively_pymdp.egg-info/top_level.txt
+pymdp/__init__.py
+pymdp/agent.py
+pymdp/control.py
+pymdp/default_models.py
+pymdp/inference.py
+pymdp/learning.py
+pymdp/maths.py
+pymdp/utils.py
+pymdp/algos/__init__.py
+pymdp/algos/fpi.py
+pymdp/algos/mmp.py
+pymdp/algos/mmp_old.py
+pymdp/envs/__init__.py
+pymdp/envs/env.py
+pymdp/envs/grid_worlds.py
+pymdp/envs/mdp_search_env.py
+pymdp/envs/tmaze.py
+pymdp/envs/visual_foraging.py
+test/test_SPM_validation.py
+test/test_agent.py
+test/test_control.py
+test/test_demos.py
+test/test_inference.py
+test/test_learning.py
+test/test_mmp.py
+test/test_wrappers.py
