### 2022-01-24 #4067 @emdupre
:wave: Hi @seankmartin, @richrobe, @semihakbayrak, @patrickmineault !

Would you be willing to review _pymdp: A Python library for active inference in discrete state spaces_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/infer-actively/pymdp) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.04067/joss.04067/10.21105.joss.04067.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you !
