### 2022-01-25 #4098 @whedon
[REVIEW]: pymdp: A Python library for active inference in discrete state spaces
**Submitting author:** @conorheins (<a href="http://orcid.org/0000-0002-5884-7728">Conor Heins</a>)
**Repository:** https://github.com/infer-actively/pymdp
**Branch with paper.md** (empty if default branch): master
**Version:** v0.0.5
