### 2022-01-25 #4098 @emdupre
Hi everyone :wave: Thanks again for agreeing to review this submission ! The review will take place in this issue.

Whenever possible, please open relevant issues on the [linked software repository ](https://github.com/infer-actively/pymdp)(and cross-link them with this issue) rather than discussing them here. This helps to make sure that feedback is translated into actionable items to improve the software.

If you aren't sure how to get started, please see the [Reviewing for JOSS guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) -- and, of course, feel free to ping me with any questions !
