### 2022-02-01 infer-actively/pymdp#67 @seankmartin
Full API documentation
Hi!

I think it would be useful to provide full API documentation for your package on your [RTD website](https://pymdp-rtd.readthedocs.io/en/latest/).
For example, I think the agent class and the envs subpackage are quite important, but I can't find API docs for them on the site.
I'd find it easier to follow if they were listed in the modules section of the site, or even if you wanted to provide a separate full API documentation section that would work too I think.
Thanks!
