### 2022-02-01 infer-actively/pymdp#68 @seankmartin
README badges
I think it would be helpful to have the badges in README to link to something other than the picture of the badge.
E.g. the pypi could link to https://pypi.org/project/inferactively-pymdp/
Further I would suggest a badge to link to your read the docs site, and if you have CI set up for your pytest tests, a badge for that would be very helpful
