### 2022-02-01 #4098 @seankmartin
I'm in progress with the review.
I've raised a non-blocking issue about some missing API documentation here https://github.com/infer-actively/pymdp/issues/67.
The paper is missing an explicit summary section, which I think should be added.
Will post a final update after checking functionality!
