### 2022-02-02 #4098 @patrickmineault
I'm going to wait until @seankmartin completes a first pass until I look at this.
