### 2022-02-16 #4098 @conorheins
> I'm in progress with the review. I've raised a non-blocking issue about some missing API documentation here [infer-actively/pymdp#67](https://github.com/infer-actively/pymdp/issues/67). The paper is missing an explicit summary section, which I think should be added. Will post a final update after checking functionality!

Hi @emdupre, I'm wondering if we can already start responding to the issues raised by @seankmartin ? Or should we wait until both reviewers are finished before we start responding? Thanks :)
