### 2022-02-17 #4098 @emdupre
Hi @conorheins, thanks for confirming on this ! It's entirely up to you whether you'd like to address @seankmartin and @patrickmineault's reviews as they come in or all at once.

After the reviews are in, I'll check-in with you on the status of your response. If you've been addressing comments as they come in, this will usually mean confirming that you're ready to receive additional reviewer feedback on your revisions. If instead you've elected to wait until all comments are available, then I'll ask for your anticipated timeline for making revisions.

If you have any other questions, please don't hesitate to ask.
