### 2022-02-24 #4098 @seankmartin
Ok I'm finished a first pass now.
Sincere apologies for the delay in particular to @patrickmineault since you were waiting for me.

I am very happy with the functionality of the package. The tests all pass and are useful, and I have manually verified some outputs in places which also were correct.
I have also run all of the demos/tutorials. I think the cue chaining (Epistemic Chaining) demo is particularly well written and is probably the best branching point for doing your own stuff with pymdp.
I like that the env and agent classes follow a similar pattern to OpenAI Gym, just makes it easier to work with.

My main concerns would be as before then (about API docs and no explicit summary section in paper). To reiterate slightly:
The package in general is very well documented, such as in the `control.py` file where the docstrings are very informative and have API docs on your read the docs website.
I think that the Agent and Env class are critical and that they should have class docstrings and should be included on the read the docs website

Overall I think the package is great and clearly has had an lot of effort put into it!
