### 2022-03-12 d371e96
Moved old `Usage` section above and renamed to `Summary` (as it was originally in arxiv version). In reference to https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1027354821
