### 2022-03-12 3d5763c
changed reference to `infer_states_test()` to `_infer_states_test` in SPM test suite, after changing that function a private member in `agent.py` (to avoid autodocumentation)
