### 2022-03-12 c1e6fc4
changed references to second return of `agent.infer_policies()` from `efe` to `G`
