### 2022-03-12 61acc10
added documentation to Sphinx autodocumentation for Agent API class, based on https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1049772353
