+++ pymdp/agent.py
@@ -67,7 +67,7 @@ class Agent(object):
         self.lr_pB = lr_pB
         self.lr_pD = lr_pD
 
-        """ Initialise observation model (A matrices) """
+        # Initialise observation model (A matrices)
         if not isinstance(A, np.ndarray):
             raise TypeError(
                 'A matrix must be a numpy array'
@@ -77,14 +77,14 @@ class Agent(object):
 
         assert utils.is_normalized(self.A), "A matrix is not normalized (i.e. A.sum(axis = 0) must all equal 1.0"
 
-        """ Determine number of observation modalities and their respective dimensions """
+        # Determine number of observation modalities and their respective dimensions
         self.num_obs = [self.A[m].shape[0] for m in range(len(self.A))]
         self.num_modalities = len(self.num_obs)
 
-        """ Assigning prior parameters on observation model (pA matrices) """
+        # Assigning prior parameters on observation model (pA matrices)
         self.pA = pA
 
-        """ Initialise transition model (B matrices) """
+        # Initialise transition model (B matrices)
         if not isinstance(B, np.ndarray):
             raise TypeError(
                 'B matrix must be a numpy array'
@@ -98,7 +98,7 @@ class Agent(object):
         self.num_states = [self.B[f].shape[0] for f in range(len(self.B))]
         self.num_factors = len(self.num_states)
 
-        """ Assigning prior parameters on transition model (pB matrices) """
+        # Assigning prior parameters on transition model (pB matrices) 
         self.pB = pB
 
         # If no `num_controls` are given, then this is inferred from the shapes of the input B matrices
@@ -130,7 +130,7 @@ class Agent(object):
 
         assert all([n_c == max_action for (n_c, max_action) in zip(self.num_controls, list(np.max(all_policies, axis =0)+1))]), "Maximum number of actions is not consistent with `num_controls`"
 
-        """ Construct prior preferences (uniform if not specified) """
+        # Construct prior preferences (uniform if not specified)
 
         if C is not None:
             if not isinstance(C, np.ndarray):
@@ -146,7 +146,7 @@ class Agent(object):
         else:
             self.C = self._construct_C_prior()
 
-        """ Construct prior over hidden states (uniform if not specified) """
+        # Construct prior over hidden states (uniform if not specified)
     
         if D is not None:
             if not isinstance(D, np.ndarray):
@@ -167,11 +167,10 @@ class Agent(object):
 
         assert utils.is_normalized(self.D), "A matrix is not normalized (i.e. A.sum(axis = 0) must all equal 1.0"
 
-        """ Assigning prior parameters on initial hidden states (pD vectors) """
+        # Assigning prior parameters on initial hidden states (pD vectors)
         self.pD = pD
 
-        """ Construct prior over policies (uniform if not specified) """
-
+        # Construct prior over policies (uniform if not specified) 
         if E is not None:
             if not isinstance(E, np.ndarray):
                 raise TypeError(
@@ -255,6 +254,20 @@ class Agent(object):
         return E
 
     def reset(self, init_qs=None):
+        """
+        Resets the posterior beliefs about hidden states of the agent to a uniform distribution, and resets time to first timestep of the simulation's temporal horizon.
+        Returns the posterior beliefs about hidden states.
+
+        Returns
+        ---------
+        qs: ``numpy.ndarray`` of dtype object
+           Initialized posterior over hidden states. Depending on the inference algorithm chosen and other parameters (such as the parameters stored within ``edge_handling_paramss),
+           the resulting ``qs`` variable will have additional sub-structure to reflect whether beliefs are additionally conditioned on timepoint and policy.
+            For example, in case the ``self.inference_algo == 'MMP' `, the indexing structure of ``qs`` is policy->timepoint-->factor, so that 
[…]
