+++ docs/agent.rst
@@ -1,0 +1,5 @@
+Agent class
+=================================
+
+.. autoclass:: pymdp.agent.Agent
+    :members:
