### 2022-03-12 aef6abf
added quick usage example and extra information to `Agent` class docstring
