### 2022-03-13 01b75dc
added env autodocumentation to `index.rst`
