### 2022-03-13 ea83bca
added reference to env.rst in index.rst
