### 2022-03-13 #4098 @conorheins
Hi @seankmartin , thanks for your comments and reviewing our package. We believe we have addressed the main API issues you raised. The latest version of the RTD page with new sections for `Agent` and `Env` API documentation can be found here: https://pymdp-rtd.readthedocs.io/en/paper_branch.

I haven't addressed your issue about the README.md badges yet: https://github.com/infer-actively/pymdp/issues/68 but I thought I'd just post an update here now, to signal that we believe we've addressed Issue https://github.com/infer-actively/pymdp/issues/67 and the main brunt of your review.

Note we've also added an explicit summary section in the paper, also as requested.
