+++ pymdp/jax/agent.py
@@ -1,0 +1,461 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+""" Agent Class
+
+__author__: Conor Heins, Alexander Tschantz, Daphne Demekas, Brennan Klein
+
+"""
+
+import warnings
+import numpy as np
+from pymdp import inference, control, learning
+from pymdp import utils, maths
+import copy
+
+class Agent(object):
+    """ 
+    The Agent class, the highest-level API that wraps together processes for action, perception, and learning under active inference.
+
+    The basic usage is as follows:
+
+    >>> my_agent = Agent(A = A, B = C, <more_params>)
+    >>> observation = env.step(initial_action)
+    >>> qs = my_agent.infer_states(observation)
+    >>> q_pi, G = my_agent.infer_policies()
+    >>> next_action = my_agent.sample_action()
+    >>> next_observation = env.step(next_action)
+
+    This represents one timestep of an active inference process. Wrapping this step in a loop with an ``Env()`` class that returns
+    observations and takes actions as inputs, would entail a dynamic agent-environment interaction.
+    """
+
+    def __init__(
+        self,
+        A,
+        B,
+        C=None,
+        D=None,
+        E = None,
+        pA=None,
+        pB = None,
+        pD = None,
+        num_controls=None,
+        policy_len=1,
+        inference_horizon=1,
+        control_fac_idx=None,
+        policies=None,
+        gamma=16.0,
+        use_utility=True,
+        use_states_info_gain=True,
+        use_param_info_gain=False,
+        action_selection="deterministic",
+        inference_algo="VANILLA",
+        inference_params=None,
+        modalities_to_learn="all",
+        lr_pA=1.0,
+        factors_to_learn="all",
+        lr_pB=1.0,
+        lr_pD=1.0,
+        use_BMA = True,
+        policy_sep_prior = False,
+        save_belief_hist = False
+    ):
+
+        ### Constant parameters ###
+
+        # policy parameters
+        self.policy_len = policy_len
+        self.gamma = gamma
+        self.action_selection = action_selection
+        self.use_utility = use_utility
+        self.use_states_info_gain = use_states_info_gain
+        self.use_param_info_gain = use_param_info_gain
+
+        # learning parameters
+        self.modalities_to_learn = modalities_to_learn
+        self.lr_pA = lr_pA
+        self.factors_to_learn = factors_to_learn
+        self.lr_pB = lr_pB
+        self.lr_pD = lr_pD
+
+        self.A = A
+
+        # self.A = pytree.map(utils.normalized, A)
+
+        """ Determine number of observation modalities and their respective dimensions """
+        self.num_obs = [self.A[m].shape[0] for m in range(len(self.A))]
+        self.num_modalities = len(self.num_obs)
+
+        """ Assigning prior parameters on observation model (pA matrices) """
+        self.pA = pA
+
+        # self.B = map( utils.normalized, B)
+        self.B = B
+
+        # Determine number of hidden state factors and their dimensionalities
+        self.num_states = [self.B[f].shape[0] for f in range(len(self.B))]
+        self.num_factors = len(self.num_states)
+
+        """ Assigning prior parameters on transition model (pB matrices) """
+        self.pB = pB
+
+        # If no `num_controls` are given, then this is inferred from the shapes of the input B matrices
+        self.num_controls = [self.B[f].shape[2] for f in range(self.num_factors)]
+
+        # Users have the option to make only certain factors controllable.
+        # default behaviour is to make all hidden state factors controllable
+        # (i.e. self.num_states == self.num_controls)
+        self.control_fac_idx = control_fac_idx
+        self.policies = policies
+
+        self.C = C
+
[…]
