### 2022-03-15 a3d23c5
first steps of implementing fixed point iteration in `jax`. log-likelihood computation using jax functions complete
