+++ pymdp/jax/algos.py
@@ -1,0 +1,8 @@
+from maths import compute_likelihood
+
+def run_vanilla_fpi(A, obs, prior):
+    """ Vanilla fixed point iteration (jaxified) """
+
+    likelihood = compute_likelihood(obs, A)
+
+    pass
