### 2022-03-28 c96d522
fpi implementation with jax
