### 2022-04-01 fcf1152
turn action into integer when indexing into slices of B array in `update_state_likelihood_dirichlet()`, to avoid improper index error
