### 2022-04-01 89284df
wrapper fn in`default_models()` library for building local-navigation transition model (B matrix) for 2d grid worlds
