### 2022-04-01 2dee7b1
compute_log_likelihood already assumes onehot observations
