### 2022-04-01 e89a2fb
unit test comparing JAX accelerated fixed-point iteration vs standard numpy version
