### 2022-04-02 42c6f73
- added reference to paper on RTD front page (to address Issue #64) - added new section in RTD for notebooks that are inspired by MATLAB code in Step-by-Step tutorial paper
