### 2022-04-05 894652c
fixed fpi and corrected numerics of the gradient
