+++ pymdp/jax/utils.py
@@ -1,0 +1,615 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+""" Utility functions
+
+__author__: Conor Heins, Alexander Tschantz, Brennan Klein
+"""
+
+import numpy as np
+
+import jax.numpy as jnp
+
+from typing import (Any, Callable, List, NamedTuple, Optional, Sequence, Union, Tuple)
+
+Tensor = Any # maybe jnp.ndarray, but typing seems not to be well defined for jax
+Vector = List[Tensor]
+Shape = Sequence[int]
+ShapeList = list[Shape]
+
+EPS_VAL = 1e-16 # global constant for use in norm_dist()
+
+# def sample(probabilities):
+#     sample_onehot = np.random.multinomial(1, probabilities.squeeze())
+#     return np.where(sample_onehot == 1)[0][0]
+
+# def sample_obj_array(arr):
+#     """ 
+#     Sample from set of Categorical distributions, stored in the sub-arrays of an object array 
+#     """
+    
+#     samples = [sample(arr_i) for arr_i in arr]
+
+#     return samples
+
+# def obj_array(num_arr):
+#     """
+#     Creates a generic object array with the desired number of sub-arrays, given by `num_arr`
+#     """
+#     return np.empty(num_arr, dtype=object)
+
+# def obj_array_zeros(shape_list):
+#     """ 
+#     Creates a numpy object array whose sub-arrays are 1-D vectors
+#     filled with zeros, with shapes given by shape_list[i]
+#     """
+#     arr = obj_array(len(shape_list))
+#     for i, shape in enumerate(shape_list):
+#         arr[i] = np.zeros(shape)
+#     return arr
+
+def list_array_uniform(shape_list: ShapeList) -> Vector:
+    """ 
+    Creates a numpy object array whose sub-arrays are uniform Categorical
+    distributions with shapes given by shape_list[i]. The shapes (elements of shape_list)
+    can either be tuples or lists.
+    """
+    arr = []
+    for shape in shape_list:
+        arr.append( norm_dist(jnp.ones(shape)) )
+    return arr
+
+# def obj_array_ones(shape_list, scale = 1.0):
+#     arr = obj_array(len(shape_list))
+#     for i, shape in enumerate(shape_list):
+#         arr[i] = scale * np.ones(shape)
+    
+#     return arr
+
+# def onehot(value, num_values):
+#     arr = np.zeros(num_values)
+#     arr[value] = 1.0
+#     return arr
+
+# def random_A_matrix(num_obs, num_states):
+#     if type(num_obs) is int:
+#         num_obs = [num_obs]
+#     if type(num_states) is int:
+#         num_states = [num_states]
+#     num_modalities = len(num_obs)
+
+#     A = obj_array(num_modalities)
+#     for modality, modality_obs in enumerate(num_obs):
+#         modality_shape = [modality_obs] + num_states
+#         modality_dist = np.random.rand(*modality_shape)
+#         A[modality] = norm_dist(modality_dist)
+#     return A
+
+# def random_B_matrix(num_states, num_controls):
+#     if type(num_states) is int:
+#         num_states = [num_states]
+#     if type(num_controls) is int:
+#         num_controls = [num_controls]
+#     num_factors = len(num_states)
+#     assert len(num_controls) == len(num_states)
+
+#     B = obj_array(num_factors)
+#     for factor in range(num_factors):
+#         factor_shape = (num_states[factor], num_states[factor], num_controls[factor])
+#         factor_dist = np.random.rand(*factor_shape)
+#         B[factor] = norm_dist(factor_dist)
+#     return B
+
+# def random_single_categorical(shape_list):
+#     """
[…]
