### 2022-04-06 infer-actively/pymdp#71 @patrickmineault
Typos in Tutorial 1
There are two instances which read *THe*, which should be changed to *The*

A heading is mislabeled:

3. The prior over observations: the D vector or P(s)

Should read:

4. The prior over hidden states: the D vector or P(s)
