### 2022-04-06 6a8e621
fix typos in the text of Tutorial 1 (mis-labelling prior over hidden states and two instances of "THe" vs. "the"), to address Issue #71: https://github.com/infer-actively/pymdp/issues/71#issue-1195197784
