### 2022-04-06 #4098 @patrickmineault
Apologies for the lateness of my review - this is great! The authors very clearly explain the need in the companion paper - currently, the mysteries of the free energy principle are accessible to few and tucked away in a corner of the Matlab package  SPM. By making their methods widely accessible across a wide audience, they're opening the field. I really appreciate the effort that went into making the API compatible with OpenAI gym - and would love to see deeper comparisons of RL and POMDP free energy agents. As a side note, the star history (+150% stars since the start of the year) makes it clear that people already find it useful.

The docs are generally great, and the emphasis on tutorials should make it easy for anyone to get up and running. I've opened a few very minor, and optional issues on the Github repo about the first two tutorials. I see that you've addressed @seankmartin 's issues with Agent being undocumented in the latest RDC, which was also a point of contention.

It's ready to ship as far as I'm concerned.
