### 2022-04-07 #4098 @conorheins
Hi @patrickmineault,

Thanks very much for the kinds words/ thoughtful reviews, particularly the attention you gave to the tutorials - those were all fantastic suggestions and I will get down to incorporating your changes this weekend at the latest.
