### 2022-04-07 #4098 @emdupre
Thank you @seankmartin and @patrickmineault for your thoughtful reviews !

@conorheins, I see that https://github.com/infer-actively/pymdp/issues/67, https://github.com/infer-actively/pymdp/issues/68, https://github.com/infer-actively/pymdp/issues/71, and https://github.com/infer-actively/pymdp/issues/72 are all open from this review process, though the reviewers are quite consistent in noting that these are optional asks.

Could you please confirm when you're finished incorporating this reviewer feedback and are ready to move forward with the submission ? I'll also provide a few editorial comments shortly.
