### 2022-04-07 #4098 @conorheins
Hi @emdupre,

Even though they're optional, I can address those remaining issues this weekend (namely https://github.com/infer-actively/pymdp/issues/71, and https://github.com/infer-actively/pymdp/issues/72) before we fully proceed -- I'd also like to edit an affiliations and a funding acknowledgement, if that's okay. As soon as I'm finished I'll post here to confirm that I'm all set.

And a quick question regarding the closing of issues -- I was planning to wait until paper_branch was merged with master before I closed https://github.com/infer-actively/pymdp/issues/67 (and the others of course, but that's the only one I've already addressed) but I can already close it now if that's preferable for the review process.

Thanks,
Conor
