### 2022-04-07 #4098 @emdupre
> As soon as I'm finished I'll post here to confirm that I'm all set.

That sounds perfect ! Thank you.

> I was planning to wait until paper_branch was merged with master before I closed https://github.com/infer-actively/pymdp/issues/67 (and the others of course, but that's the only one I've already addressed) but I can already close it now if that's preferable for the review process. Would that be preferable?

It makes it a bit clearer that you consider the issue(s) entirely address, though closing before merging in the branch is not a requirement -- especially since these are optional additions to the project !

I also have a few requests on the paper itself. If you could please:
- [ ] add in DOIs to those references noted with "DOI missing" in the most [recent editorialbot comment](https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1092161263). The bot has suggested a potential DOI for each reference, but please confirm if these are correct !
- [ ] update the ForneyLab citation to match [their citation.bib file](https://github.com/biaslab/ForneyLab.jl/blob/master/CITATION.bib) ?
- [ ] standardize the formatting of arXiv references; e.g., Baioumy et al (2021) is cited with `arXiv. <arXiv-URL>` while Laar et al (2021) is cited with `arXiv Preprint <arXiv:URI>`

More generally, too: Could you clarify the relationship between this paper and [the companion arXiv preprint](https://arxiv.org/abs/2201.03904) ? There are partially overlapping sections, and I want to make sure I understand the relationship between these two resources !
