### 2022-04-09 a6729a4
changes to main Agent demo to address textual edits raised in #72 and https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1090878602
