### 2022-04-09 infer-actively/pymdp#72 @conorheins
Hi @patrickmineault thanks a lot for these helpful suggestions. I've addressed points 1-4.

That bit you discovered in Point 4 with the reward manipulation was a mistake on my part, I didn't mean to have the user flip the 'valence' of reward and punishment like that, although the outcome you observed makes perfect sense. Because in that sense, the 'meaning' of the reward and loss observations are simply switched, in terms of the agent's prior over observations (they now "expect themselves" to see Loss more than Reward, meaning the observation formerly referred to as "Loss" is technically the reward now). I've now changed that section of the demo to make the agent more risk-averse, but still not Loss-seeking.

I'm going to leave this issue open now, because I haven't added in text to address your last point, which is also a great suggestion and should be incorporated: "I wondered whether the inference was equivalent to a finite horizon dynamic programming solution, or similar in spirit to an agent that uses a UCB heuristic or Thompson sampling. If you could have a few references in the tutorial about this, it would be great." This is related to this paper I think: https://arxiv.org/abs/2009.08111.

I will have to do some more research on these exact relationships and then incorporate that into the text of that demo.
