### 2022-04-09 infer-actively/pymdp#67 @conorheins
Thanks again for these great suggestions @seankmartin -- we've already addressed them in the review a few weeks back, but I'm just commenting again to confirm that we've closed this issue now
