### 2022-04-09 infer-actively/pymdp#71 @conorheins
Fixed :)
