### 2022-04-09 74f47b3
to address JOSS editor comments here: https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1092181486 - added missing DOIs for particular citations - used official ForneyLab citation from their GitHub page -standardized ArXiv-preprint reference style
