+++ paper/paper.bib
@@ -31,7 +31,8 @@
   title={Chance-Constrained Active Inference},
   author={van de Laar, Thijs and Senoz, Ismail and {\"O}z{\c{c}}elikkale, Ay{\c{c}}a and Wymeersch, Henk},
   journal={arXiv preprint arXiv:2102.08792},
-  year={2021}
+  year={2021},
+  do = {10.1162/neco_a_01427}
 }
 
 @Article{         harris2020array,
@@ -58,11 +59,10 @@
 }
 
 @article{millidge2021understanding,
-    title = {Understanding the origin of information-seeking exploration in probabilistic objectives for control},
-    author = {Millidge, Beren and Tschantz, Alexander and Seth, Anil and Buckley, Christopher},
-    journal = {arXiv},
-    year = {2021},
-    url = {https://arxiv.org/abs/2103.06859}
+  title={Understanding the origin of information-seeking exploration in probabilistic objectives for control},
+  author={Millidge, Beren and Tschantz, Alexander and Seth, Anil and Buckley, Christopher},
+  journal={arXiv preprint arXiv:2103.06859},
+  year={2021}
 }
 
 @article{millidge2021whence,
@@ -438,12 +438,15 @@
     doi = {10.1503/jpn.200032}
 }
 
-@article{smith2021step,
-    title = {A step-by-step tutorial on active inference and its application to empirical data},
-    author = {Smith, Ryan and Friston, Karl J. and Whyte, Christopher},
-    journal = {PsyArXiv},
-    year = {2021},
-    doi = {10.31234/osf.io/b4jm6}
+@article{smith2022step,
+  title={A step-by-step tutorial on active inference and its application to empirical data},
+  author={Smith, Ryan and Friston, Karl J and Whyte, Christopher J},
+  journal={Journal of Mathematical Psychology},
+  volume={107},
+  pages={102632},
+  year={2022},
+  publisher={Elsevier},
+  doi={10.1016/j.jmp.2021.102632}
 }
 
 @article{friston2008variational,
@@ -458,11 +461,10 @@
 }
 
 @article{friston2019free,
-    title = {A free energy principle for a particular physics},
-    author = {Friston, Karl J.},
-    journal = {arXiv},
-    year = {2019},
-    url = {https://arxiv.org/abs/1906.10184}
+  title={A free energy principle for a particular physics},
+  author={Friston, Karl},
+  journal={arXiv preprint arXiv:1906.10184},
+  year={2019}
 }
 
 @article{millidge2020deep,
@@ -485,19 +487,28 @@
     doi = {10.1007/978-3-030-64919-7_1}
 }
 
-@inproceedings{van2018forneylab,
-    title = {ForneyLab: A toolbox for biologically plausible free energy minimization in dynamic neural models},
-    author = {van de Laar, Thijs and Cox, Marco and Senoz, Ismail and Bocharov, Ivan and de Vries, Bert},
-    booktitle = {Conference on Complex Systems},
-    year = {2018},
-    url = {https://biaslab.github.io/pdf/forneylab/van_de_laar_forneylab_biologically_plausible_free_energy_minimization.pdf}
+@article{ForneyLab2019,
+	title = {A factor graph approach to automated design of {Bayesian} signal processing algorithms},
+	volume = {104},
+	issn = {0888-613X},
+	url = {http://www.sciencedirect.com/science/article/pii/S0888613X18304298},
+	doi = {10.1016/j.ijar.2018.11.002},
[…]
