### 2022-04-09 589f796
updated with new string for ForneyLab citation, changed Thomas' acknowledgement to be explicit about finding technical error in arxiv version
