### 2022-04-09 #4098 @conorheins
Hi @emdupre,

Thanks for pointing out these issues, I've attempted to address them all. Let me know if there are any further issues. For the missing DOIs, some of the suggestions were correct, some of them were not, so I just manually added them all in. I hope I did so correctly :)

- [x] add in DOIs to those references noted with "DOI missing" in the most https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1092161263. The bot has suggested a potential DOI for each reference, but please confirm if these are correct

- [x] update the ForneyLab citation to match [their citation.bib file](https://github.com/biaslab/ForneyLab.jl/blob/master/CITATION.bib) ?

- [x] standardize the formatting of arXiv references; e.g., Baioumy et al (2021) is cited with arXiv. <arXiv-URL> while Laar et al (2021) is cited with arXiv Preprint <arXiv:URI>

And finally, regarding the clarification between preprint companion and this JOSS submission -- we wrote the longer, technical version first, to be out there for those who are interested in the mathematical background of these models  (POMDPs) that one can build in `pymdp`. Then we decided to make the submission for JOSS  just a shortened, more cleaned up / tighter version of the arXiv version, that focuses on the software itself and its goals/advantages. That's why you'll notice groups of overlapping text (particularly in the earlier sections), because for all intents and purposes the arXiv preprint version is just a longer, more technical version of the JOSS submission.

So given all that: would it be okay for us to keep the arXiv preprint online, as the more technical version of the JOSS article, but then ask users to officially cite the JOSS version? I think it would be helpful to have that longer, more mathematical version with the appendices available as a pedagogical resource. But obviously we would want the JOSS version to be the "official" version that is cited.
