### 2022-04-09 #4098 @conorheins
> Hi @emdupre,
>
> Even though they're optional, I can address those remaining issues this weekend (namely [infer-actively/pymdp#71](https://github.com/infer-actively/pymdp/issues/71), and [infer-actively/pymdp#72](https://github.com/infer-actively/pymdp/issues/72)) before we fully proceed -- I'd also like to edit an affiliations and a funding acknowledgement, if that's okay. As soon as I'm finished I'll post here to confirm that I'm all set.
>
> And a quick question regarding the closing of issues -- I was planning to wait until paper_branch was merged with master before I closed [infer-actively/pymdp#67](https://github.com/infer-actively/pymdp/issues/67) (and the others of course, but that's the only one I've already addressed) but I can already close it now if that's preferable for the review process. Would that be preferable?
>
> Thanks, Conor

Oh and I forgot to confirm the resolution of the issues. Regarding the optional requests raised by the reviewers, I've now addressed nearly all of them, and just left some minor bits open that I'd like to work out in the future. But otherwise I am ready to proceed with the submission.
