### 2022-04-09 b41ddf6
missing comma after publisher in gregory 1980 citation
