+++ paper/paper.bib
@@ -519,7 +519,7 @@
   number={1038},
   pages={181--197},
   year={1980},
-  publisher={The Royal Society London}
+  publisher={The Royal Society London},
   doi={10.1098/rstb.1980.0090}
 }
 
