### 2022-04-11 #4098 @emdupre
Thank you, @conorheins !

I'm confirming the best course of action under JOSS policy for the arXiv companion piece, and I will have an answer for you -- along with next steps -- shortly.
