### 2022-04-11 9da57b5
fixed typo in doi of van2021chance reference (see https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1094442176)
