+++ paper/paper.bib
@@ -32,7 +32,7 @@
   author={van de Laar, Thijs and Senoz, Ismail and {\"O}z{\c{c}}elikkale, Ay{\c{c}}a and Wymeersch, Henk},
   journal={arXiv preprint arXiv:2102.08792},
   year={2021},
-  do = {10.1162/neco_a_01427}
+  doi = {10.1162/neco_a_01427}
 }
 
 @Article{         harris2020array,
