### 2022-04-18 #4098 @emdupre
Hi @conorheins ,

Thank you for your patience on this ! I've consulted with other JOSS editors and have outlined a list of next steps that I hope will be helpful. I'm detailing them below, and please do let me know if you have any questions in addressing them.

- [ ] It would be best to directly cite the arXiv companion piece in the JOSS article such that the relationship between the two is clearly documented in the publication. This will help readers to consult the appropriate resource, as necessary, and maintain a record of both pieces. I would suggest including this citation in the `Statement of Need` section to expand on your stated goal of increasing the "accessibility and exposure of the active inference framework to researchers, engineers, and developers with diverse disciplinary backgrounds."
- [ ] Thank you for standardizing the arXiv references ! I received a bit more information on the best practice in formatting these. Could you please modify them to include the `eprint`, `archivePrefix`, and `primaryClass` fields? So, for example, for [millidge2021understanding](https://github.com/infer-actively/pymdp/blob/9da57b57683bae1be65db8145b0c50c4862388ce/paper/paper.bib#L61-L66) the reference would be:
  ```
  @misc{millidge2021understanding,
  title={Understanding the origin of information-seeking exploration in probabilistic objectives for control},
  author={Millidge, Beren and Tschantz, Alexander and Seth, Anil and Buckley, Christopher},
  eprint={2103.06859v7},
  archivePrefix={arXiv},
  primaryClass={cs.LG},
  year={2021}
  }
    ```
 - [ ] For [`van2021chance`](https://github.com/infer-actively/pymdp/blob/9da57b57683bae1be65db8145b0c50c4862388ce/paper/paper.bib#L30-L35), could you update the reference from the arXiv preprint to the published _Neural Computation_ article:
   ```
   @article{10.1162/neco_a_01427,
    author={van de Laar, Thijs and Senoz, Ismail and {\"O}z{\c{c}}elikkale, Ay{\c{c}}a and Wymeersch, Henk},
    title = {Chance-Constrained Active Inference},
    journal = {Neural Computation},
    volume = {33},
    number = {10},
    pages = {2710-2735},
    year = {2021},
    month = {09},
    issn = {0899-7667},
    doi = {10.1162/neco_a_01427},
    url = {https://doi.org/10.1162/neco\_a\_01427},
    eprint = {https://direct.mit.edu/neco/article-pdf/33/10/2710/1963501/neco\_a\_01427.pdf},
    }
   ```

Pending these changes, could you then please:

- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata (e.g., the JOSS paper and software archive have the same title and author listing)
- [ ] List the DOI of the archived version here.

I can then move forward with accepting the submission 🚀
