### 2022-04-18 09acaac
updated chance-constrained AIF reference/citation to the official NeuralComp one suggested by @emdupre here: https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1101083477
