+++ paper/paper.bib
@@ -27,13 +27,20 @@
   organization={Springer}
 }
 
-@article{van2021chance,
-  title={Chance-Constrained Active Inference},
-  author={van de Laar, Thijs and Senoz, Ismail and {\"O}z{\c{c}}elikkale, Ay{\c{c}}a and Wymeersch, Henk},
-  journal={arXiv preprint arXiv:2102.08792},
-  year={2021},
-  doi = {10.1162/neco_a_01427}
-}
+@article{10.1162/neco_a_01427,
+ author={van de Laar, Thijs and Senoz, Ismail and {\"O}z{\c{c}}elikkale, Ay{\c{c}}a and Wymeersch, Henk},
+ title = {Chance-Constrained Active Inference},
+ journal = {Neural Computation},
+ volume = {33},
+ number = {10},
+ pages = {2710-2735},
+ year = {2021},
+ month = {09},
+ issn = {0899-7667},
+ doi = {10.1162/neco_a_01427},
+ url = {https://doi.org/10.1162/neco\_a\_01427},
+ eprint = {https://direct.mit.edu/neco/article-pdf/33/10/2710/1963501/neco\_a\_01427.pdf},
+ }
 
 @Article{         harris2020array,
  title         = {Array programming with {NumPy}},
