### 2022-04-25 302a89a
standardized arxiv references to include fields requested in  https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1101083477
