+++ paper/paper.bib
@@ -65,11 +65,13 @@
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
 
-@article{millidge2021understanding,
-  title={Understanding the origin of information-seeking exploration in probabilistic objectives for control},
-  author={Millidge, Beren and Tschantz, Alexander and Seth, Anil and Buckley, Christopher},
-  journal={arXiv preprint arXiv:2103.06859},
-  year={2021}
+@misc{millidge2021understanding,
+    title={Understanding the origin of information-seeking exploration in probabilistic objectives for control},
+    author={Millidge, Beren and Tschantz, Alexander and Seth, Anil and Buckley, Christopher},
+    eprint={2103.06859v7},
+    archivePrefix={arXiv},
+    primaryClass={cs.LG}, 
+    year={2021}
 }
 
 @article{millidge2021whence,
@@ -467,10 +469,12 @@
     doi = {10.1016/j.neuroimage.2008.02.054}
 }
 
-@article{friston2019free,
+@misc{friston2019free,
   title={A free energy principle for a particular physics},
   author={Friston, Karl},
-  journal={arXiv preprint arXiv:1906.10184},
+  eprint={1906.10184v1},
+  archivePrefix={arXiv},
+  primaryClass={q-bio.NC}, 
   year={2019}
 }
 
@@ -510,12 +514,13 @@
 	pages = {185--204}
 }
 
-@article{baltieri2019generative,
+@misc{baltieri2019generative,
   title={Generative models as parsimonious descriptions of sensorimotor loops},
   author={Baltieri, Manuel and Buckley, Christopher L},
-  journal={arXiv preprint arXiv:1904.12937},
-  year={2019},
-  doi={10.1017/S0140525X19001353}
+  eprint={1904.12937v1},
+  archivePrefix={arXiv},
+  primaryClass={q-bio.NC}, 
+  year={2019}
 }
 
 @article{gregory1980perceptions,
@@ -530,10 +535,12 @@
   doi={10.1098/rstb.1980.0090}
 }
 
-@article{brockman2016openai,
+@misc{brockman2016openai,
   title={Openai gym},
   author={Brockman, Greg and Cheung, Vicki and Pettersson, Ludwig and Schneider, Jonas and Schulman, John and Tang, Jie and Zaremba, Wojciech},
-  journal={arXiv preprint arXiv:1606.01540},
+  eprint={1606.01540v1},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG},
   year={2016}
 }
 
@@ -576,13 +583,13 @@
     url = {http://www.cs.toronto.edu/~hinton/absps/optimal.pdf}
 }
 
-@article{baioumy2021towards,
+@misc{baioumy2021towards,
     title = {Towards stochastic fault-tolerant control using precision learning and active inference},
     author = {Baioumy, Mohamed and Pezzato, Corrado and Corbato, Carlos Hernandez and Hawes, Nick and Ferrari, Riccardo},
-    journal={arXiv preprint arXiv:2109.05870},
-    year = {2021},
-    url = {https://arxiv.org/abs/2109.05870},
-    doi = {10.1007/978-3-030-93736-2_48}
+    eprint={2109.05870v1},
+    archivePrefix={arXiv},
+    primaryClass={cs.RO},
+    year={2021}
 }
 
 @article{ccatal2021robot,
