### 2022-04-25 ff7450c
added reference to arxiv preprint and github repo in the statement of need section, following suggestion from https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1101083477
