+++ paper/paper.bib
@@ -65,6 +65,15 @@
  url           = {https://doi.org/10.1038/s41586-020-2649-2}
 }
 
+@misc{heins2022pymdp_arxiv},
+    title={pymdp: A Python library for active inference in discrete state spaces},
+    author={Heins, Conor and Millidge, Beren and Demekas, Daphne and Klein, Brennan and Friston, Karl and Couzin, Iain and Tschantz, Alexander},
+    eprint={2201.03904v1},
+    archivePrefix={arXiv},
+    primaryClass={cs.AI}, 
+    year={2022}
+}
+
 @misc{millidge2021understanding,
     title={Understanding the origin of information-seeking exploration in probabilistic objectives for control},
     author={Millidge, Beren and Tschantz, Alexander and Seth, Anil and Buckley, Christopher},
