### 2022-04-25 #4098 @conorheins
Hi @emdupre , I've now finished the re-formatting steps and changes to Statement of Need that you suggested.

Before I move onto the second set of steps (e.g. archiving the reviewed software in Zenodo, ... .), I just had a clarifying question about the order in which I do these steps.

The first step is "Make a tagged release of your software, and list the version tag of the archived version here".

Does this "archived version" refer to the archived version that will be generated in the subsequent steps (with Zenodo, etc.), or something else that I'm not understanding? This is probably obvious for someone who has worked with Zenodo or similar services but this is new for me, so before proceeding I just want to know what exactly I should be listing in the version tag of the tagged release.
