### 2022-04-25 #4098 @emdupre
Hi @conorheins,

Thank you for addressing these points ! There's a typo [here](https://github.com/infer-actively/pymdp/blob/137cec4fff77ea29ac9a922351d7671d79299823/paper/paper.bib#L68) that's preventing PDF compilation. If you can remove the extra `}` character, we should be able to re-generate the PDF !

> Does this "archived version" refer to the archived version that will be generated in the subsequent steps (with Zenodo, etc.), or something else that I'm not understanding? This is probably obvious for someone who has worked with Zenodo or similar services but this is new for me, so before proceeding I just want to know what exactly I should be listing in the version tag of the tagged release.

This is a good question; thank you for confirming ! What I mean is that when you create a new release of your software, this release is the one that is then archived in subsequent steps (e.g. on Zenodo using [this workflow](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content)), as you suggested. The motivation is that you have [a tagged release](https://docs.github.com/en/repositories/releasing-projects-on-github/managing-releases-in-a-repository) that includes the state of the software as it is described in the JOSS paper, and by archiving this tagged release it will have a persistent identifier; e.g. a DOI, that we can include in the JOSS publication.

Hopefully that helps to answer your questions, but please let me know if there's anything else I can clarify !
