### 2022-04-25 ebc0cbf
fixed typo in pymdp arxiv citation
