### 2022-04-25 #4098 @conorheins
Hi @emdupre, thanks for the clarification.

As requested, here is a link to the latest release: [v0.0.5](https://github.com/infer-actively/pymdp/tree/v0.0.5)

And here is the DOI (from Zenodo): 10.5281/zenodo.6484849

I have also verified that the Zenodo archive version has the correct metadata (title and author list).
