### 2022-04-25 be2d428
Update README.md
Added documentation and DOI badges to README.md
