### 2022-04-25 8232fae
changed references to docs to `stable` version because `latest` version is not showing correct front page
