### 2022-04-25 #4098 @emdupre
Thank you, @conorheins !

A few remaining points:
- [ ] The [software archive](https://doi.org/10.5281/zenodo.6484849) title is currently `infer-actively/pymdp: 0.0.5` -- could you please update it to `pymdp: A Python library for active inference in discrete
state spaces` to match the JOSS paper ?
- [ ] Similarly, the [software archive](https://doi.org/10.5281/zenodo.6484849) author listing is ` conorheins; Alexander Tschantz; Arun-Niranjan; Beren; mahault` when I would expect it to be `Conor Heins, Beren Millidge, Daphne Demekas, Brennan Klein, Karl Friston, Iain D. Couzin, and Alexander Tschantz`. Could you please update the author formatting to match the the formatting included in the paper (e.g., changing `conorheins` to `Conor Heins`) and confirm that you intended to have a different author list between the software paper and archive ?
- [ ] For the references that editorialbot identifies as "missing," I believe that you have included the `baltieri` reference twice in your bib list, with the second `baltieri2019pid` reference corresponding to the later version of the same article. `baioumy2021towards` also has a published version, but I assume you intended to keep the arXiv version listed s.t. readers can access the non-paywalled version. Please let me know if either of these assessments are incorrect.

These first two edits can be made directly on Zenodo by editing the metadata associated with the archive, as [described here](https://help.zenodo.org/#technical) !
