### 2022-04-26 #4098 @conorheins
Hi @emdupre,

Re: software archive metadata. Sorry about that, that's strange. I changed those metadata (title / author list) to the correct versions after I made the Zenodo release, but it seems it doesn't show up for the public-facing version, presumably because I didn't click "Publish" yet, only "Save". I've now clicked "Publish" so all the metadata should be consistent now. Let me know if you still notice any discrepancies. The title and author list should be identical to those on the paper now, hopefully.

re: Baltieri referrences -- `baltieri2019pid` (an entropy paper) is actually a different paper than `baltieri2019` (a commentary on another paper), so I think that should be fine.

re: `baouimy2021towards` -- you are correct that I intended to keep the arXiv version listed.
