### 2022-04-27 #4098 @emdupre
Great, thank you ! Only the `@baltieri2019pid` citation is currently included in the paper.md, so I wanted to confirm their relationship. I think, though, that having an uncited reference in the .bib file is not against JOSS policy, so we can move forward as-is.

I'm now happy to recommend pymdp to the EIC team for publication, and I just want to add my congratulations to you on such an impressive effort !
