### 2022-04-27 #4098 @conorheins
Hi @emdupre,

Awesome, thank you! This is great news.

Before the paper is published, would now be a good time for me to just pass it around the co-authors to make sure everyone's name / affiliations are listed properly? Or is there an equivalent of a 'final proofs' stage coming later, at which small edits like that would be appropriate?

Thanks,
Conor
