### 2022-04-28 #4098 @emdupre
When you're updating affiliations, @conorheins, can you please also:

 - [ ] Remove the uncited `baltieri2019generative` reference from the .bib file
 - [ ] Update the `baouimy2021towards` reference to the paywalled version, doi: 10.1007/978-3-030-93736-2_48

I received additional guidance that we will need the the version-of-record in the JOSS publication.
