### 2022-04-28 d2b1d2b
- removed older baltieri reference from `paper.bib` - updated baouimy2021towards citation to be book chapter version see https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1111636124
