+++ paper/paper.bib
@@ -523,15 +523,6 @@
 	pages = {185--204}
 }
 
-@misc{baltieri2019generative,
-  title={Generative models as parsimonious descriptions of sensorimotor loops},
-  author={Baltieri, Manuel and Buckley, Christopher L},
-  eprint={1904.12937v1},
-  archivePrefix={arXiv},
-  primaryClass={q-bio.NC}, 
-  year={2019}
-}
-
 @article{gregory1980perceptions,
   title={Perceptions as hypotheses},
   author={Gregory, Richard Langton},
@@ -592,13 +583,15 @@
     url = {http://www.cs.toronto.edu/~hinton/absps/optimal.pdf}
 }
 
-@misc{baioumy2021towards,
-    title = {Towards stochastic fault-tolerant control using precision learning and active inference},
-    author = {Baioumy, Mohamed and Pezzato, Corrado and Corbato, Carlos Hernandez and Hawes, Nick and Ferrari, Riccardo},
-    eprint={2109.05870v1},
-    archivePrefix={arXiv},
-    primaryClass={cs.RO},
-    year={2021}
+@inproceedings{baioumy2021towards,
+    title={Towards Stochastic Fault-Tolerant Control Using Precision Learning and Active Inference},
+    author={Baioumy, Mohamed and Pezzato, Corrado and Corbato, Carlos Hernandez and Hawes, Nick and Ferrari, Riccardo},
+    booktitle={Machine Learning and Principles and Practice of Knowledge Discovery in Databases},
+    year={2022},
+    publisher={Springer International Publishing},
+    pages={681--691},
+    isbn={978-3-030-93736-2},
+    doi = {10.1007/978-3-030-93736-2_48}
 }
 
 @article{ccatal2021robot,
