### 2022-04-29 8d5a4aa
jitable agent class and response likelihood
