+++ pymdp/jax/control.py
@@ -3,35 +3,35 @@
 # pylint: disable=no-member
 # pylint: disable=not-an-iterable
 
-import itertools
 import jax.numpy as jnp
 from functools import partial
-from jax import lax, vmap, nn
-from maths import *
+from jax import lax, jit, vmap, nn
 from itertools import chain
+
+from pymdp.jax.maths import *
 # import pymdp.jax.utils as utils
 
-def update_posterior_policies(policy_matrix, qs_init, A, B, log_C, gamma = 16.0):
+
+def update_posterior_policies(policy_matrix, qs_init, A, B, C, gamma = 16.0):
     # policy --> n_levels_factor_f x 1
     # factor --> n_levels_factor_f x n_policies
     ## vmap across policies
-    compute_G_fixed_states = partial(compute_G_policy, qs_init, A, B, log_C)
+    compute_G_fixed_states = partial(compute_G_policy, qs_init, A, B, C)
 
     # only in the case of policy-dependent qs_inits
     # in_axes_list = (1,) * n_factors
     # all_efe_of_policies = vmap(compute_G_policy, in_axes=(in_axes_list, 0))(qs_init_pi, policy_matrix)
 
-    neg_efe_all_policies = vmap(compute_G_fixed_states)(policy_matrix)
     # policies needs to be an NDarray of shape (n_policies, n_timepoints, n_control_factors)
-
-    # @TODO: convert negative EFE of each policy into a posterior probability
+    neg_efe_all_policies = vmap(compute_G_fixed_states)(policy_matrix)
 
     return nn.softmax(gamma * neg_efe_all_policies), neg_efe_all_policies
 
 def compute_expected_state(qs_prior, B, u_t): 
     """
     Compute posterior over next state, given belief about previous state, transition model and action...
-    """  
+    """
+    assert len(u_t) == len(B)  
     qs_next = []
     for qs_f, B_f, u_f in zip(qs_prior, B, u_t):
         qs_next.append( B_f[..., u_f].dot(qs_f) )
@@ -59,33 +59,11 @@ def factor_dot(A, qs):
 
     return res
 
-def factor_dot_2(A, qs):
-    """ Dot product of a multidimensional array with `x`.
-    
-    Parameters
-    ----------
-    - `x` [1D numpy.ndarray] - either vector or array of arrays
-        The alternative array to perform the dot product with
-    
-    Returns 
-    -------
-    - `Y` [1D numpy.ndarray] - the result of the dot product
-    """
-
-    x = qs[0]
-    for q in qs[1:]:
-        x = jnp.expand_dims(x, -1) * q
-
-    joint = A * x
-    dim = joint.shape[0]
-    return joint.reshape(dim, -1).sum(-1)
-
 def compute_expected_obs(qs, A):
 
     qo = []
     for A_m in A:
         qo.append( factor_dot(A_m, qs) )
-        # qo.append( factor_dot_2(A_m, qs) )
 
     return qo
 
@@ -103,11 +81,11 @@ def compute_info_gain(qs, qo, A):
     
     return H_qo - (qs_H_A * x).sum()
     
-def compute_expected_utility(qo, log_C):
+def compute_expected_utility(qo, C):
     
     util = 0.
-    for o_m, log_C_m in zip(qo, log_C):
-        util += (o_m * log_C_m).sum()
+    for o_m, C_m in zip(qo, C):
+        util += (o_m * C_m).sum()
     
     return util
 
@@ -118,6 +96,7 @@ def compute_G_policy(qs_init, A, B, C, policy_i):
[…]
