### 2022-05-02 #4098 @conorheins
Hi @emdupre, I've gotten confirmation that affiliations/names are all correct :) I've also updated .bib file as requested
