### 2022-05-02 #4098 @emdupre
Thank you, @conorheins ! I have one last reference request, since the included Hipólito reference was published in _Frontiers_ over the weekend. Can you please update [this reference](https://github.com/infer-actively/pymdp/blob/d2b1d2b5cae4b9af9255d821a30f39e4be8dfd89/paper/paper.bib#L646-L651) to:

```
@article{hipolito2021enactive,
author = {Hip{\'o}lito, In{\^e}s and van Es, Thomas},
title = {Enactive-Dynamic Social Cognition and Active Inference},
journal = {Frontiers in Psychology},
volume = {13},
year = {2022},
url = {https://www.frontiersin.org/article/10.3389/fpsyg.2022.855074},
doi = {10.3389/fpsyg.2022.855074},
issn = {1664-1078}
}
```
