+++ paper/paper.bib
@@ -42,29 +42,6 @@
  eprint = {https://direct.mit.edu/neco/article-pdf/33/10/2710/1963501/neco\_a\_01427.pdf},
  }
 
-@Article{         harris2020array,
- title         = {Array programming with {NumPy}},
- author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
-                 van der Walt and Ralf Gommers and Pauli Virtanen and David
-                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
-                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
-                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
-                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
-                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
-                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
-                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
-                 Travis E. Oliphant},
- year          = {2020},
- month         = sep,
- journal       = {Nature},
- volume        = {585},
- number        = {7825},
- pages         = {357--362},
- doi           = {10.1038/s41586-020-2649-2},
- publisher     = {Springer Science and Business Media {LLC}},
- url           = {https://doi.org/10.1038/s41586-020-2649-2}
-}
-
 @misc{heins2022pymdp_arxiv,
     title={pymdp: A Python library for active inference in discrete state spaces},
     author={Heins, Conor and Millidge, Beren and Demekas, Daphne and Klein, Brennan and Friston, Karl and Couzin, Iain and Tschantz, Alexander},
@@ -72,15 +49,6 @@
     archivePrefix={arXiv},
     primaryClass={cs.AI}, 
     year={2022}
-}
-
-@misc{millidge2021understanding,
-    title={Understanding the origin of information-seeking exploration in probabilistic objectives for control},
-    author={Millidge, Beren and Tschantz, Alexander and Seth, Anil and Buckley, Christopher},
-    eprint={2103.06859v7},
-    archivePrefix={arXiv},
-    primaryClass={cs.LG}, 
-    year={2021}
 }
 
 @article{millidge2021whence,
@@ -116,88 +84,6 @@
     doi = {10.1038/s41598-018-38246-3}
 }
 
-@book{wainwright2008graphical,
-    title = {Graphical models, exponential families, and variational inference},
-    author = {Wainwright, Martin J. and Jordan, Michael Irwin},
-    year = {2008},
-    volume = {1},
-    publisher = {Now Publishers Inc},
-    doi = {10.1561/2200000001}
-}
-
-@book{beal2003variational,
-    title = {{Variational Algorithms for Approximate Bayesian Inference}},
-    author = {Beal, Matthew James},
-    year = {2003},
-    publisher = {University of London, University College London},
-    url = {https://cse.buffalo.edu/faculty/mbeal/thesis/}
-}
-
-@inproceedings{yedidia2000generalized,
-    title = {Generalized belief propagation},
-    author = {Yedidia, Jonathan S. and Freeman, William T. and Weiss, Yair},
-    booktitle = {Advances in Neural Information Processing Systems},
-    volume = {13},
-    pages = {689--695},
-    year = {2000},
-    url = {https://dl.acm.org/doi/10.5555/3008751.3008848}
-}
-
-@article{winn2005variational,
-    title = {Variational message passing},
-    author = {Winn, John and Bishop, Christopher M.},
[…]
