### 2022-05-02 #4098 @conorheins
Hi @emdupre , I removed that reference as well as several others that weren't used here (some of them were left over from the arxiv version). Thanks
