### 2022-05-04 #4098 @kyleniemeyer
Hi @conorheins, I'm the AEIC on duty this week doing some final checks.

In your article, can you remove the explicit arXiv URL from the end of the statement of need section? The citation you included is sufficient and includes the URL.

Similarly, the article already links to the software repository, so I think you should remove the URL pointing to that in the same paragraph. If you really want to keep a link in the text, perhaps just make the word "repository" a hyperlink like `[repository](https://github.com/infer-actively/pymdp)`.

In the Introduction first paragraph, can you wrap "e.g." with commas, like "... in, e.g., psychophysics ...". Similarly, please add a comma after "e.g." in later uses.

It looks like the Parr et al. 2020 article from _Cerebral Cortex_ is missing a DOI, can you check if one is available?
