### 2022-05-04 9c26b9e
-removed arxiv URL from statement of need section -removed github URL from statement of need section -wrapped e.g.'s in commas -added doi to parr2020prefrontal citation
In response to: https://github.com/openjournals/joss-reviews/issues/4098#issuecomment-1117613827
