+++ paper/paper.bib
@@ -574,7 +574,8 @@
   number={2},
   pages={682--695},
   year={2020},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/cercor/bhz118}
 }
 
 @article{smith2021long,
