### 2022-05-04 #4098 @kyleniemeyer
Congratulations @conorheins on your article's publication in JOSS! ([Conor Heins](http://orcid.org/0000-0002-5884-7728))
